+++ HomeCinema/frmLoading.cs
@@ -1,5 +1,4 @@
 ﻿using HomeCinema.Global;
-using Microsoft.Build.Evaluation;
 using System;
 using System.Drawing;
 using System.Windows.Forms;
@@ -8,41 +7,32 @@ namespace HomeCinema
 {
     public partial class frmLoading : Form
     {
-        public frmLoading()
+        public frmLoading(Form parent)
         {
             InitializeComponent();
             Icon = new Icon(GlobalVars.FILE_ICON);
-            FormClosing += new FormClosingEventHandler(frmLoading_FormClosing);
+            //Show(parent);
+            // Load image from app path and assign to picBox
+            picBox.Image = GlobalVars.IMG_LOADING;
         }
 
         private void frmLoading_Load(object sender, EventArgs e)
         {
-            // at form load
-            string file = GlobalVars.PATH_RES + "Loading.gif";
-            picBox.Image = Image.FromFile(file);
-        }
+            // Load image from app path and assign to picBox
+            if (picBox.Image == null)
+            {
+                picBox.Image = GlobalVars.IMG_LOADING;
+            }
+            // Center on screen
+            CenterToParent();
+            /*
+            Rectangle r = new Rectangle();
+            r = PARENT.RectangleToScreen(PARENT.ClientRectangle);
 
-        // Dispose Image loading
-        private void frmLoading_FormClosing(object sender, FormClosingEventArgs e)
-        {
-            if (picBox.Image != null)
-            {
-                picBox.Image.Dispose();
-            }
-            foreach (Control c in Controls)
-            {
-                c.Dispose();
-            }
-            //Dispose();
-        }
-
-        private void timeClose_Tick(object sender, EventArgs e)
-        {
-            Form fc = Application.OpenForms["frmMain"];
-            if (fc == null)
-            {
-                Close();
-            }
+            int x = r.Left + ((r.Width - Width) / 2);
+            int y = r.Top + ((r.Height - Height) / 2);
+            Location = new Point(x, y);
+            */
         }
     }
 }