+++ VERSION_HISTORY.md
@@ -2,11 +2,16 @@
 
 ## HomeCinema
 
-****v0.0.0.1 build c0b7b7d** - (19 April 2020 PHT) First Public Release of HomeCinema**
+****v0.0.0.3 build a57d298** - (30 April 2020 PHT) GUI Update**
 
 - **Links:**
-  -  **[Windows 32/64 bit](https://github.com/JerloPH/HomeCinema/releases/download/v0.0.0.1/HomeCinema-Windows_v0.0.0.1.zip)**
+  -  **[Windows 32/64 bit](link_here)**
 	
+- **What's New?**
+  - Added new ["File"](/Resources/media_ext.hc_data "media_ext.hc_data") that loads supported extensions for media files.
+  - Changed GUI.
+  - Changed "No Cover" Image file.
+
 ****v0.0.0.2 build a57d298** - (26 April 2020 PHT) Second Release - First Update**
 
 - **Links:**
@@ -19,4 +24,11 @@
   - Movie / TV Show item Container is bigger
   - Add year to title of Movie / TV Show
   - Cleaned the GUI
+
+****v0.0.0.1 build c0b7b7d** - (19 April 2020 PHT) First Public Release of HomeCinema**
+
+- **Links:**
+  -  **[Windows 32/64 bit](https://github.com/JerloPH/HomeCinema/releases/download/v0.0.0.1/HomeCinema-Windows_v0.0.0.1.zip)**
+	
+